rock tile fix, pause menu button size change, pause menu doesn't pause now (weird with it being multiplayer)

+++ Source/Client/Unit.cpp
@@ -75,7 +75,7 @@ void Unit::update(float dt)
 				else if (selected)
 				{
 					std::cout << "unit deselected" << mouse_x << ", " << mouse_y << "\n";
-					moveToPosition(mouse_x, mouse_y);
+					moveToPosition((float)mouse_x, (float)mouse_y);
 					serializePacketType = PacketType::MOVE;
 					sendPacket();
 					selected = false;
